refactor: move common types and constants to subfolder
in preparation for a smaller, more compact Footer variant
