+++ frontend/src/components/Footer/common/OgpFullLogo.tsx
@@ -1,0 +1,5 @@
+import { chakra } from '@chakra-ui/react'
+
+import { ReactComponent as OgpFullLogoSvg } from '~assets/svgs/brand/ogp-logo-full.svg'
+
+export const OgpFullLogo = chakra(OgpFullLogoSvg)
